relocate display_summary to shared location
